engine: keep just one audio beat tracker
