+++ engine/audio/src/CMakeLists.txt
@@ -9,8 +9,6 @@ add_library(${module_name}
     audioplugincache.cpp audioplugincache.h
     audiorenderer.cpp audiorenderer.h
     beattracker.cpp beattracker.h
-    beattrackerauto.cpp beattrackerauto.h
-    beattracking.cpp beattracking.h
 )
 set_property(TARGET ${module_name} PROPERTY POSITION_INDEPENDENT_CODE ON)
 target_include_directories(${module_name} PUBLIC
