+++ engine/audio/src/audiocapture.cpp
@@ -24,13 +24,7 @@
 #include <qmath.h>
 
 #include "audiocapture.h"
-#if defined(AUTO_TRACKER)
-  #include "beattrackerauto.h"
-#elif defined(NEW_TRACKER)
-  #include "beattracker.h"
-#else
-  #include "beattracking.h"
-#endif
+#include "beattracker.h"
 
 #define USE_HANNING
 #define CLEAR_FFT_NOISE
@@ -80,16 +74,7 @@ AudioCapture::AudioCapture (QObject* parent)
     m_plan_forward = fftw_plan_dft_r2c_1d(m_bufferSize, m_fftInputBuffer,
                                           reinterpret_cast<fftw_complex*>(m_fftOutputBuffer), 0);
 #endif
- #if defined(AUTO_TRACKER)
-    m_beatTracker = new BeatTrackerAuto(m_sampleRate, m_channels);
- #elif defined(NEW_TRACKER)
-    m_beatTracker = new BeatTracker(m_sampleRate, m_bufferSize, m_channels, 86, 1.3);
-    m_beatTracker->setBand(40.0, 400.0);      // bit wider band for now
-    m_beatTracker->setFluxSmoothing(0.6);     // less smoothing
-    m_beatTracker->setMinBeatInterval(0.20);  // ~300 BPM max
- #else
-    m_beatTracker = new BeatTracking(m_sampleRate, m_channels);
- #endif
+    m_beatTracker = new BeatTracker(m_sampleRate, m_channels);
 }
 
 AudioCapture::~AudioCapture()
@@ -364,15 +349,7 @@ void AudioCapture::run()
                 processData();
 
                 if (m_beatTracker->processAudio(m_audioBuffer, m_captureSize))
-                {
-#if defined(AUTO_TRACKER)
                     emit beatDetected(qRound(m_beatTracker->bpm()));
-#elif defined(NEW_TRACKER)
-                    emit beatDetected(qRound(m_beatTracker->getCurrentBpm()));
-#else
-                    emit beatDetected(qRound(m_beatTracker->currentBpm()));
-#endif
-                }
             }
             else
             {
