+++ engine/test/CMakeLists.txt
@@ -1,6 +1,6 @@
 project(test)
 
-add_subdirectory(beattrackerauto)
+add_subdirectory(beattracker)
 add_subdirectory(beattrackerbench)
 add_subdirectory(bus)
 add_subdirectory(channelsgroup)
