+++ engine/test/beattracker/test.sh
@@ -1,4 +1,4 @@
 #!/bin/sh
 export LD_LIBRARY_PATH=../../src
 export DYLD_FALLBACK_LIBRARY_PATH=../../src
-./beattrackerauto_test
+./beattracker_test
